Publish the conformer and transition-state upload contracts
`POST /api/v1/uploads/conformers` and
`POST /api/v1/uploads/transition-states` are live, non-deprecated routes,
and until now their request bodies lived inside the server. A consumer
could not pin them, and a breaking change to either shipped with no
signal a client could act on. The ARC adapter alone targets these two
contracts with 188 of its 775 mappings.

Both request models, and the closure they reach that was not already on
the wire, now live in `tckdb-schemas`:

- `TransitionStateUploadRequest` (+ `TSReactionUpload`,
  `TSReactionParticipantUpload`) — every dependency was already
  published, so this is a straight move.
- `ConformerUploadRequest` (+ `ConformerUploadStatmechPayload`,
  `ElectronicLevelIn`), which pulled two things across with it:
  `TransportUploadPayload`, and the statmech create-side payloads
  (`StatmechSourceCalculationCreate`, `StatmechTorsionCreate`,
  `StatmechTorsionCoordinateCreate`) together with the `*Base` classes
  that define their fields. The backend's `*Read`/`*Update` schemas keep
  inheriting from those bases, so the field definitions still have one
  home; only the create side is on the wire.

`tckdb_schemas.workflows` now re-exports every published request model,
so `from tckdb_schemas.workflows import ConformerUploadRequest` works.

Nothing was inverted to get here: the boundary test needed no
relaxation, because the dependency closure of both models was already
either published or backend-side only by accident of where the file sat.
The backend modules become re-export shims of the same class objects, so
route signatures are untouched and the generated OpenAPI is
byte-identical — `tests/api/golden/openapi.json` needed no regeneration.

The claim is tested rather than asserted.
`tests/api/test_wire_only_upload_contracts.py` builds both payloads in a
subprocess where `app` is genuinely unimportable, then posts them to the
live routes. Getting a backend-free interpreter out of a dev environment
takes `-S`: `tckdb-backend` is editable-installed here, and an editable
install is reachable through a `.pth` finder that no `PYTHONPATH` can
hide. The boundary test's runtime check now covers the new modules too.
