+++ schemas/python/tckdb-schemas/tckdb_schemas/workflows/__init__.py
@@ -1,0 +1,41 @@
+"""Published upload request contracts.
+
+Every name here is a top-level request body accepted by a live route, or a
+payload nested directly inside one. Importing from this package is the
+supported way for a consumer to pin an upload contract:
+
+    from tckdb_schemas.workflows import ConformerUploadRequest
+
+Nothing under ``app`` is reachable from here — see
+``tests/test_import_boundaries.py``.
+"""
+
+from tckdb_schemas.workflows.computed_reaction_upload import (
+    ComputedReactionUploadRequest,
+)
+from tckdb_schemas.workflows.computed_species_upload import (
+    ComputedSpeciesUploadRequest,
+)
+from tckdb_schemas.workflows.conformer_upload import (
+    ConformerUploadRequest,
+    ConformerUploadStatmechPayload,
+    ElectronicLevelIn,
+)
+from tckdb_schemas.workflows.transition_state_upload import (
+    TransitionStateUploadRequest,
+    TSReactionParticipantUpload,
+    TSReactionUpload,
+)
+from tckdb_schemas.workflows.transport_upload import TransportUploadPayload
+
+__all__ = [
+    "ComputedReactionUploadRequest",
+    "ComputedSpeciesUploadRequest",
+    "ConformerUploadRequest",
+    "ConformerUploadStatmechPayload",
+    "ElectronicLevelIn",
+    "TSReactionParticipantUpload",
+    "TSReactionUpload",
+    "TransitionStateUploadRequest",
+    "TransportUploadPayload",
+]
